Fixed the boolean problem with the threads in controller.
The Listener instantiates the controller class (and USB) on every
actionevent, as this code is in the listener.

Since we assigned the values false and false in the constructor, the
values would change back to false even if the variables were static
(same regardless of the instance the class is in).

Moving the assignment in the variable declaration fixes this.

Interesting enough, the 3 values are assigned to -1 in the constructor
as well, but we get to assign them to a different value afterwards in
the code.